requirements.txt
google-generativeai>=0.3.0
chromadb>=0.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.5

